requirements.txt
pandas
plotly
requests
psycopg2-binary
flask
dash
python-dotenv
matplotlib
